+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.18] - 2020-11-16
+
+### Changed
+- added option `forceDoorState` for `PLC_GarageDoorOpener`
+- Home app seems does not to use the lock for `PLC_GarageDoorOpener`. Remove `get_LockCurrentState`, `get_LockTargetState` and `set_LockTargetState`
+  
 ## [1.0.17] - 2020-11-14
 
 ### Changed